packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Himansh Mudigonda", email = "himudigonda@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0",
    "passlib[argon2]>=1.7.4",
]

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"
